signup api change for guidelines step, start implementing
