+++ app/frontend/src/features/auth/jail/constants.ts
@@ -1,29 +1,3 @@
-import { ReactComponent as ghost } from "./resources/ghost.svg";
-import { ReactComponent as handshake } from "./resources/handshake.svg";
-import { ReactComponent as handWithHeart } from "./resources/handWithHeart.svg";
-import { ReactComponent as shield } from "./resources/shield.svg";
-
-export const COMMUNITY_GUIDELINES = [
-  "Couchers.org is an inclusive community where people of all cultures and backgrounds come together. Treat everyone with mutual respect and dignity",
-  "Couchers.org is NOT for dating or hooking up. Harassment or or using the platform for the purpose of dating or hooking up will result in a ban.",
-  "Read profiles, references, and resources. Report any inappropriate content or behavior.",
-  "Keep things non-transactional. Don't exchange money while staying or hosting. Couchers.org is about learning, experiences, and getting to know people.",
-];
-export const COMMUNITY_GUIDELINE_ICONS = [
-  handWithHeart,
-  ghost,
-  shield,
-  handshake,
-];
-export const COMMUNITY_GUIDELINE_LABEL = "Ok, got it";
-export const COMMUNITY_GUIDELINE_TITLES = [
-  "Be kind to one another",
-  "This is not a dating app",
-  "Be safe and sensible",
-  "Keep money out of it",
-];
-export const COMMUNITY_GUIDELINES_SECTION_HEADING = "Community Guidelines";
-export const COMMUNITY_GUIDELINES_REQUIRED = "All checkboxes are required";
 export const JAIL_TOS_TEXT =
   "We've update our Terms of Service. To continue, please read and accept the new ";
 export const LOCATION_SECTION_HEADING = "Please add your location";
