+++ app/frontend/src/features/auth/signup/CommunityGuidelinesForm.tsx
@@ -1,0 +1,32 @@
+import Alert from "components/Alert";
+import { useAuthContext } from "features/auth/AuthProvider";
+import { Error as GrpcError } from "grpc-web";
+import { useMutation } from "react-query";
+import { service } from "service";
+
+export default function CommunityGuidelinesForm() {
+  const { authActions } = useAuthContext();
+
+  const mutation = useMutation<void, GrpcError, boolean>(
+    async (accept) => {
+      const state = await service.auth.signupFlowCommunityGuidelines(accept);
+      authActions.updateSignupState(state);
+    },
+    {
+      onMutate() {
+        authActions.clearError();
+      },
+      onSettled() {
+        window.scroll({ top: 0, behavior: "smooth" });
+      },
+    }
+  );
+
+  return (
+    <>
+      {mutation.error && (
+        <Alert severity="error">{mutation.error.message || ""}</Alert>
+      )}
+    </>
+  );
+}
