+++ app/frontend/src/service/auth.ts
@@ -1,3 +1,4 @@
+import { BoolValue } from "google-protobuf/google/protobuf/wrappers_pb";
 import { HostingStatus } from "proto/api_pb";
 import {
   ContributorForm as ContributorFormPb,
@@ -95,6 +96,13 @@ export async function signupFlowFeedback(
 export async function signupFlowEmailToken(emailToken: string) {
   const req = new SignupFlowReq();
   req.setEmailToken(emailToken);
+  const res = await client.auth.signupFlow(req);
+  return res.toObject();
+}
+
+export async function signupFlowCommunityGuidelines(accept: boolean) {
+  const req = new SignupFlowReq();
+  req.setAcceptCommunityGuidelines(new BoolValue().setValue(accept));
   const res = await client.auth.signupFlow(req);
   return res.toObject();
 }
